+++ app/components/comparison/components/comparison-data.service.ts
@@ -2,6 +2,9 @@ import { Injectable, ChangeDetectorRef } from '@angular/core';
 import { Http } from '@angular/http';
 import { TableDataSet, Data, Property, ListItem, RatingSet } from './../shared/index';
 import { ComparisonService } from './comparison.service';
+import { LocalStorageService } from "angular-2-local-storage";
+
+declare const moment;
 
 @Injectable()
 export class ComparisonDataService {
@@ -9,14 +12,16 @@ export class ComparisonDataService {
     private tags: {[name: string]: string; } = {};
 
     constructor(private http: Http,
-                private comparisonService: ComparisonService) {
+                private comparisonService: ComparisonService,
+                private lss: LocalStorageService) {
     }
 
     public loadData(tableDataSet: TableDataSet, cd: ChangeDetectorRef) {
+        const self = this;
         this.http.request('app/components/comparison/data/data.json')
             .subscribe(res => {
                 res.json().forEach(obj => {
-                    const data: Data = new Data(this.http);
+                    const data: Data = new Data(this.lss, this, this.comparisonService);
                     data.tag = obj.tag;
                     const regArray =
                         /^((?:(?:\w+\s*)(?:-?\s*\w+.)*)+)\s*-?\s*((?:(?:http|ftp|https)(?::\/\/)(?:[\w_-]+(?:(?:\.[\w_-]+)+))|(?:www.))(?:[\w.,@?^=%&:\/~+#-]*[\w@?^=%&\/~+#-])?)$/gi
@@ -71,6 +76,25 @@ export class ComparisonDataService {
             });
     }
 
+    public getRepoData(data: Data, repo: string) {
+        repo = repo.replace(/^-\s*/, "");
+        this.http.get(this.repoQueryBuildUrl(repo)).toPromise().then(function (res) {
+            const body = JSON.parse(res["_body"]);
+            const date = moment(body[0].commit.author.date);
+            const sync = moment();
+            data.setRepoData({lastCommit: date.toDate(), lastSync: sync.toDate()});
+        });
+    }
+
+    private repoQueryBuildUrl(repoUrl: string) {
+        let url: string;
+        if (/https?:\/\/github\.com.*/.test(repoUrl.trim())) {
+            url = repoUrl.trim().replace(/https?:\/\/github.com/, "https://api.github.com/repos");
+            url += url.endsWith("/") ? "commits" : "/commits";
+        }
+        return url;
+    }
+
     public getDefaultAttachmentTags(): Array<string> {
         const tags: Array<string> = [];
         for (const key in this.tags) {
@@ -84,5 +108,9 @@ export class ComparisonDataService {
 
     public getLength(): number {
         return this.data.length;
+    }
+
+    public getData(): Array<Data> {
+        return this.data;
     }
 }