+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Enabling and disabling elements
+- Improved foreground colors when auto assigning
 
 ## [1.0.0] - 2017-08-03
 ### Added